Strip comments from color schemes
This has been a problem.  Because xml.parser.expat doesn't seem to be
able to handle comments like this: `<!-- comments --- -->`, it is
probably best to strip them out.

Another common problem is that people put comments before the encoding
declaration.  This will address that as well.  It is easier to just
clean it up than educate people.

ref: #24 and #22